Reducing warnings on Windows

+++ Opcodes/sequencer.c
@@ -126,7 +126,7 @@ static int32_t sequencer(CSOUND *csound, SEQ *p)
     if (*p->verbos) printf("RESET!!\n");
     goto minus7;
   }
-  else if (p->time > CS_KSMPS) {         /* Not yet time to act */
+  else if (p->time > (int32_t) CS_KSMPS) {         /* Not yet time to act */
     //printf("**time= %d", p->time);
     p->time -= CS_KSMPS;
     *p->res = -FL(1.0);
